+++ lib/liquex/filter.ex
@@ -224,9 +224,9 @@ defmodule Liquex.Filter do
       iex> Liquex.Filter.date("March 14, 2016", "%b %d, %y", %{})
       "Mar 14, 16"
   """
-  def date(%Date{} = value, format, _), do: Timex.format!(value, format, :strftime)
-  def date(%DateTime{} = value, format, _), do: Timex.format!(value, format, :strftime)
-  def date(%NaiveDateTime{} = value, format, _), do: Timex.format!(value, format, :strftime)
+  def date(%Date{} = value, format, _), do: Calendar.strftime(value, format)
+  def date(%DateTime{} = value, format, _), do: Calendar.strftime(value, format)
+  def date(%NaiveDateTime{} = value, format, _), do: Calendar.strftime(value, format)
 
   def date("now", format, context), do: date(DateTime.utc_now(), format, context)
   def date("today", format, context), do: date(Date.utc_today(), format, context)