+++ lib/liquex/tag/raw_tag.ex
@@ -1,0 +1,24 @@
+defmodule Liquex.Tag.RawTag do
+  @moduledoc false
+
+  @behaviour Liquex.Tag
+  import NimbleParsec
+
+  alias Liquex.Parser.Tag
+
+  def parse do
+    endraw = Tag.tag_directive("endraw")
+
+    text =
+      lookahead_not(endraw)
+      |> utf8_char([])
+      |> times(min: 1)
+      |> reduce({Kernel, :to_string, []})
+
+    ignore(Tag.tag_directive("raw"))
+    |> optional(text)
+    |> ignore(endraw)
+  end
+
+  def render(contents, context), do: {contents, context}
+end